feat: move general/log path to log
